// ombi-api/model/ombi-core-models-search-search-movie-view-model.ts
export interface OmbiCoreModelsSearchSearchMovieViewModel {
    /**
     * 
     * @type {boolean}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    adult?: boolean;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    backdropPath?: string;
    /**
     * 
     * @type {Array<number>}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    genreIds?: Array<number>;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    originalLanguage?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    originalTitle?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    overview?: string;
    /**
     * 
     * @type {number}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    popularity?: number;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    posterPath?: string;
    /**
     * 
     * @type {Date}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    releaseDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    title?: string;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    video?: boolean;
    /**
     * 
     * @type {number}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    voteAverage?: number;
    /**
     * 
     * @type {number}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    voteCount?: number;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    alreadyInCp?: boolean;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    trailer?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    homepage?: string;
    /**
     * 
     * @type {number}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    rootPathOverride?: number;
    /**
     * 
     * @type {number}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    qualityOverride?: number;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    type?: OmbiCoreModelsSearchSearchMovieViewModelTypeEnum;
    /**
     * 
     * @type {OmbiApiTheMovieDbModelsReleaseDatesDto}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    releaseDates?: OmbiApiTheMovieDbModelsReleaseDatesDto;
    /**
     * 
     * @type {Date}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    digitalReleaseDate?: Date;
    /**
     * 
     * @type {number}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    id?: number;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    approved?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    requested?: boolean;
    /**
     * 
     * @type {number}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    requestId?: number;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    available?: boolean;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    plexUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    embyUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    quality?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    imdbId?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    theTvDbId?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    theMovieDbId?: string;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    subscribed?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiCoreModelsSearchSearchMovieViewModel
     */
    showSubscribe?: boolean;
}

/**
    * @export
    * @enum {string}
    */
export enum OmbiCoreModelsSearchSearchMovieViewModelTypeEnum {
    TvShow = 'TvShow',
    Movie = 'Movie',
    Album = 'Album'
}